Updated the documentation to support srcSet
Added srcSet to the list of possible attributes and linked to MDN's responsive images guide. 
Expanded the already existing example to show the correct use of  the attribute.

p.s. @silencesys if you think that the srcSet example that I added clutters too much the existing code, I can write a different section with a more in depth explanation :)